upload cluster configuration
