+++ cli/spark/endpoints/cluster/cluster_create.py
@@ -95,7 +95,7 @@ def execute(args: typing.NamedTuple):
     else:
         cluster_conf.user_configuration = None
 
-    print_cluster_conf(cluster_conf, wait)
+    utils.print_cluster_conf(cluster_conf, wait)
     spinner = utils.Spinner()
     spinner.start()
 
@@ -112,30 +112,3 @@ def execute(args: typing.NamedTuple):
     else:
         log.info("Cluster %s is being provisioned.", cluster.id)
 
-
-def print_cluster_conf(cluster_conf: ClusterConfiguration, wait: bool):
-    user_configuration = cluster_conf.user_configuration
-
-    log.info("-------------------------------------------")
-    log.info("spark cluster id:        %s", cluster_conf.cluster_id)
-    log.info("spark cluster size:      %s",
-             cluster_conf.vm_count + cluster_conf.vm_low_pri_count)
-    log.info(">        dedicated:      %s", cluster_conf.vm_count)
-    log.info(">     low priority:      %s", cluster_conf.vm_low_pri_count)
-    log.info("spark cluster vm size:   %s", cluster_conf.vm_size)
-    log.info("custom scripts:          %s", cluster_conf.custom_scripts)
-    log.info("subnet ID:               %s", cluster_conf.subnet_id)
-    log.info("file shares:             %s", len(cluster_conf.file_shares) if cluster_conf.file_shares is not None else 0)
-    log.info("docker repo name:        %s", cluster_conf.docker_repo)
-    log.info("wait for cluster:        %s", wait)
-    log.info("username:                %s", user_configuration.username)
-    if user_configuration.password:
-        log.info("Password: %s", '*' * len(user_configuration.password))
-    log.info("Plugins:")
-    if len(cluster_conf.plugins) == 0:
-        log.info("    None Configured")
-    else:
-        for plugin in cluster_conf.plugins:
-            args = ", ".join([ "{0}: {1}".format(k, v) for [k, v] in plugin.args.items()])
-            log.info("  - %s: (%s)", plugin.name, args)
-    log.info("-------------------------------------------")
