+++ cli/spark/endpoints/cluster/cluster_get.py
@@ -1,6 +1,7 @@
 import argparse
 import typing
 import aztk
+from cli import log
 from cli import utils, config
 
 
@@ -9,6 +10,10 @@ def setup_parser(parser: argparse.ArgumentParser):
                         dest='cluster_id',
                         required=True,
                         help='The unique id of your spark cluster')
+    parser.add_argument('--show-config',
+                        dest='show_config',
+                        action='store_true',
+                        help='Show the cluster configuration')
 
 
 def execute(args: typing.NamedTuple):
@@ -16,3 +21,8 @@ def execute(args: typing.NamedTuple):
     cluster_id = args.cluster_id
     cluster = spark_client.get_cluster(cluster_id)
     utils.print_cluster(spark_client, cluster)
+
+    if cluster.configuration and args.show_config:
+        log.info("-------------------------------------------")
+        log.info("Cluster configuration:")
+        utils.print_cluster_conf(cluster.configuration, False)
