+++ cli/utils.py
@@ -9,6 +9,7 @@ import azure.batch.models as batch_models
 import aztk.spark
 from aztk import error
 from aztk.utils import get_ssh_key
+from aztk.models import ClusterConfiguration
 from . import log
 
 
@@ -404,3 +405,31 @@ class Spinner:
 
 def utc_to_local(utc_dt):
     return utc_dt.replace(tzinfo=datetime.timezone.utc).astimezone(tz=None).strftime("%H:%M%p %d/%m/%y")
+
+
+def print_cluster_conf(cluster_conf: ClusterConfiguration, wait: bool):
+    user_configuration = cluster_conf.user_configuration
+
+    log.info("-------------------------------------------")
+    log.info("spark cluster id:        %s", cluster_conf.cluster_id)
+    log.info("spark cluster size:      %s",
+             cluster_conf.vm_count + cluster_conf.vm_low_pri_count)
+    log.info(">        dedicated:      %s", cluster_conf.vm_count)
+    log.info(">     low priority:      %s", cluster_conf.vm_low_pri_count)
+    log.info("spark cluster vm size:   %s", cluster_conf.vm_size)
+    log.info("custom scripts:          %s", cluster_conf.custom_scripts)
+    log.info("subnet ID:               %s", cluster_conf.subnet_id)
+    log.info("file shares:             %s", len(cluster_conf.file_shares) if cluster_conf.file_shares is not None else 0)
+    log.info("docker repo name:        %s", cluster_conf.docker_repo)
+    log.info("wait for cluster:        %s", wait)
+    log.info("username:                %s", user_configuration.username)
+    if user_configuration.password:
+        log.info("Password: %s", '*' * len(user_configuration.password))
+    log.info("Plugins:")
+    if len(cluster_conf.plugins) == 0:
+        log.info("    None Configured")
+    else:
+        for plugin in cluster_conf.plugins:
+            args = ", ".join([ "{0}: {1}".format(k, v) for [k, v] in plugin.args.items()])
+            log.info("  - %s: (%s)", plugin.name, args)
+    log.info("-------------------------------------------")
